fix(client): Return nil if endpoint is empty string
Why:
Passing an empty string as an endpoint to Client when instantiating a
new client might seem like something that should never happen but I
managed to trigger it while parsing some input files to register feeds
in bulk.

What:
Check if the endpoint passed to NewClient is "" and then return
immediately nil.

Arguably, the function should instead be altered to return an error, but
that would change the signature of the function, introducing a backwards
incompatible change and break the API