client: Return nil Client if no endpoint is passed
Why:

Passing an empty string as an endpoint to the client when instantiating
a new client might seem like something that should never happen but I
managed to trigger it while parsing some input files to register feeds
in bulk.

What:
Check if the endpoint passed to NewClient is "" and then return
immediately nil.

Arguably, I should instead alter the function to return an error, but
that would change the function of the signature introduce a backwards
incompatible change.